+++ extensions/chroma/types/environment.ts
@@ -4,5 +4,6 @@ export interface IChromaEnvironmentPlacement {
    scale?: Vector3;
    position?: Vector3;
    rotation?: Vector3;
+   track?: string;
    type?: number;
 }