+++ extensions/stats/groupEvent.ts
@@ -18,23 +18,23 @@ export function countEbg(
    const ebgCount: ICountEventBoxGroup = {};
    for (let i = commonEvent.length - 1; i >= 0; i--) {
       ebgCount[commonEvent[i]] = {
-         total: 0,
-         eventBox: 0,
-         base: 0,
+         groups: 0,
+         boxes: 0,
+         bases: 0,
       };
    }
 
    for (let i = ebg.length - 1; i >= 0; i--) {
       if (!ebgCount[ebg[i].id]) {
          ebgCount[ebg[i].id] = {
-            total: 0,
-            eventBox: 0,
-            base: 0,
+            groups: 0,
+            boxes: 0,
+            bases: 0,
          };
       }
-      ebgCount[ebg[i].id].total++;
-      ebgCount[ebg[i].id].eventBox += ebg[i].boxes.length;
-      ebgCount[ebg[i].id].base += ebg[i].boxes.reduce((t, e) => t + e.events.length, 0);
+      ebgCount[ebg[i].id].groups++;
+      ebgCount[ebg[i].id].boxes += ebg[i].boxes.length;
+      ebgCount[ebg[i].id].bases += ebg[i].boxes.reduce((t, e) => t + e.events.length, 0);
    }
    return ebgCount;
 }