+++ types/beatmap/shared/environment.ts
@@ -35,10 +35,11 @@ export type EnvironmentV3Name =
    | 'RockMixtapeEnvironment'
    | 'Dragons2Environment'
    | 'Panic2Environment'
-   | 'QueenEnvironment';
+   | 'QueenEnvironment'
+   | 'LinkinPark2Environment';
 
 /** List of available 360 environment in base game. */
-export type Environment360Name = 'GlassDesertEnvironment';
+export type Environment360Name = 'GlassDesertEnvironment' | 'MultiplayerEnvironment';
 
 /** List of all available environment in base game. */
 export type EnvironmentAllName = EnvironmentName | EnvironmentV3Name | Environment360Name;